input files support
restored reading from /in directory xml & .xm.zip files. Reads are
sequential for multiple files.
