+++ app/naprrql/public/js/naprr_ui.js
@@ -231,102 +231,6 @@ function initSchoolChooserHandlerQL() {
 
 }
 
-// 
-// listen for changes to the chosen school,
-// once selected pull down all datasets
-// 
-function initSchoolChooserHandler() {
-
-    // 
-    // background load all datasets when user selects a new school
-    // 
-    $("input.autocomplete").on("change", function(event) {
-        // get the current selected school
-        var selection = $('input.autocomplete').val();
-        // console.log(selection);
-
-        var delimiter_pos = selection.indexOf(" - ");
-        // console.log(delimiter_pos);
-
-        // make sure we ignore false selections
-        if (delimiter_pos < 0) {
-            return
-        }
-
-        // strip out the acara id
-        var acaraid = selection.slice(0, delimiter_pos);
-        // console.log(">" + acaraid + "<");
-        currentASLId = acaraid; // store globally for reuse
-
-        // fetch the score summary first as is smallest dataset
-        $.get("/naprr/scoresummary/" + acaraid, function(data, status) {
-            scoresummaryData = [];
-            scoresummaryData = data;
-            console.log("score summary data downloaded for " + acaraid +
-                " elements: " + scoresummaryData.length);
-
-            if (debug) {
-                console.log(scoresummaryData);
-            }
-
-            // display on screen while other reports download
-            hideReport();
-            createScoreSummaryReport();
-            showReport();
-        });
-
-        // get domain scores
-        $.get("/naprr/domainscores/" + acaraid, function(data, status) {
-            domainscoresData = [];
-            domainscoresData = data;
-            console.log("domain scores data downloaded for " + acaraid +
-                " elements: " + domainscoresData.length);
-
-            if (debug) {
-                console.log(domainscoresData);
-            }
-
-        });
-
-        $.get("/naprr/participation/" + acaraid, function(data, status) {
-            participationData = [];
-            participationData = data;
-            console.log("participation data downloaded for " + acaraid +
-                " elements: " + participationData.length);
-
-            if (debug) {
-                console.log(participationData);
-            }
-
-        });
-
-        $.get("/naprr/schoolinfo/" + acaraid, function(data, status) {
-            schoolinfoData = {};
-            schoolinfoData = data;
-            console.log("school info data downloaded for " + acaraid);
-
-            if (debug) {
-                console.log(schoolinfoData);
-            }
-
-
-        });
-
-        $.get("/naprr/codeframe", function(data, status) {
-            codeframeData = {};
-            codeframeData = data;
-            console.log("codeframe data downloaded.");
-
-            if (debug) {
-                console.log(codeframeData);
-            }
-
-        });
-
-    });
-
-}
-
 // show the extended data modal form
 function showExtendedData() {
     $('#ed-modal').modal('open');
